Refactored article and note publishing into helpers
no-issue

This is a tidying that will allow other steps to reuse this logic.
